MAINT: _lib: move elementwise algorithm framework
