+++ scipy/_lib/meson.build
@@ -106,6 +106,7 @@ python_sources = [
   '_ccallback.py',
   '_disjoint_set.py',
   '_docscrape.py',
+  '_elementwise_algorithm.py',
   '_finite_differences.py',
   '_gcutils.py',
   '_pep440.py',
